+"Look-ahead types" (tank, infantry, engineer)
This corrects an issue where tanks, infantry and engineers could miss or get hung up on certain targets - notably, turrets - if the turret is almost or on a zone boundary. By adding `xLookAhead`, the front of the tank / ground unit will cross into the zone early enough that nearby detection should now work as expected.