Fix #7: Always add touch + mouse event listeners.
Address an issue where some laptops (which are not considered "mobile" for the purpose of this game,) but which have a touch-enabled screen, were not getting mouse input handlers assigned because of "touch support" being present via `navigator.maxTouchPoints > 0` - e.g.: Dell XPS 13.

Ultimately, attempting to detect touch support is an anti-pattern - and this is a case-in-point example of it.

As a result, the `isMobile` check is now more explicit. It determines if touch controls are shown on-screen, and enables display logic for portrait vs. landscape orientations etc. In "desktop" mode, there are no touch control overlays and "responsive" / breakpoint UI logic is greatly simplified.