SPRITE UPDATE: "Remastered" 8-bit sprites extracted from AA V1.1 for Mac.
Worth noting: It appears there is a shift in gamma, or brightness overall; everything is a bit more dark relative to my earlier pull of the Mac sprites. The Mac versions are also more colourful and detailed than the PC-DOS version, which uses 4-bit / 16 colors and is much more limited.